+++ src/main/java/com/example/desafio/itau/controllers/doc/TransactionControllerSwagger.java
@@ -40,5 +40,5 @@ public interface TransactionControllerSwagger {
             description = "Estatísticas recuperadas",
             content = @Content(schema = @Schema(implementation = StatisticModel.class))
     )
-    ResponseEntity<StatisticModel> getStatistic(int seconds);
+    ResponseEntity<StatisticModel> getStatistic(long seconds);
 }
